pyproject.toml
[project]
name = "lava"
version = "0.1.0"
description = "LAVA — LLM-Assisted Visual Analytics for Big Data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Data Engine
    "duckdb>=1.0",
    "pyarrow>=15.0",
    "psutil",
    "numba",
    # LLM Pipeline
    "openai>=1.0",
    "pydantic>=2.0",
    # Viz Renderer
    "plotly>=5.20",
    "dash>=2.17",
    "dash-bootstrap-components>=1.5",
    "datashader>=0.16",
    "colorcet",
    "pillow",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff.lint.per-file-ignores]
"main.py" = ["E402"]
